调整锁定分红操作

+++ users/models.py
@@ -352,8 +352,8 @@ class UserCoinLockLog(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     user_coin_lock = models.ForeignKey(UserCoinLock, on_delete=models.DO_NOTHING)
     coin_lock_days = models.IntegerField(verbose_name="锁定天数", default=0)
-    amount = models.DecimalField(verbose_name="锁定金额", max_digits=10, decimal_places=10,
-                                 default=0.0000000000)
+    # amount = models.DecimalField(verbose_name="锁定金额", max_digits=10, decimal_places=10, default=0.0000000000)
+    amount = models.IntegerField(verbose_name="锁定金额", default=0)
     start_time = models.DateTimeField(verbose_name="锁定开始时间")
     end_time = models.DateTimeField(verbose_name="锁定结束时间")
     created_at = models.DateTimeField(verbose_name="创建时间", auto_now_add=True)
